Code adaptation "Backward simplication"
Code adaptation "Backward simplication"
On top of the usual conditions
a) some opponent in front (but no lever)
b) some neighbours (in front) (but no neighbour behind or same rank)
c) < rank_5

to find out if a pawn is backward we look at the squares in front of
this pawn to reach the same rank as the next neighbour.

In current master, a pawn is backward if any of those squares is
controlled by an enemy pawn on an adjacent file

In this version, a pawn is ALSO backward if any of those squares is
occupied by an enemy pawn.

STC:
http://tests.stockfishchess.org/tests/view/56fe7efd0ebc59301a3541f1
LLR: 2.95 (-2.94,2.94) [-3.00,1.00]
Total: 19051 W: 3557 L: 3433 D: 12061

LTC:
http://tests.stockfishchess.org/tests/view/56febc2d0ebc59301a354209
LLR: 2.95 (-2.94,2.94) [-3.00,1.00]
Total: 40810 W: 5619 L: 5526 D: 29665

Bench: 7525245

Resolves #614
master
1 parent 8fb45ca commit 29b5842da8d5477c0aea924cfd364c9e619456a2
@Rocky640 Rocky640 committed with zamar