Fix multiPV issue #502
Fix multiPV issue #502
In general, this patch handles the cases where we don't have a valid
score for each PV line in a multiPV search. This can happen if the
search has been stopped in an unfortunate moment while still in the
aspiration loop. The patch consists of two parts.

Part 1: The new PVIdx was already part of the k-best pv's in the last
iteration, and we therefore have a valid pv and score to output from the
last iteration. This is taken care of with:

bool updated = (i <= PVIdx && rootMoves[i].score != -VALUE_INFINITE);

Case 2: The new PVIdx was NOT part of the k-best pv's in the last
iteration, and we have no valid pv and score to output. Not from the
current nor from the previous iteration. To avoid this, we are now also
considering the previous score when sorting, so that the PV lines with
no actual but with a valid previous score are pushed up again, and the
previous score can be displayed.

bool operator<(const RootMove& m) const {
return m.score != score ? m.score < score : m.previousScore <
previousScore; } // Descending sort

I also added an assertion in UCI::value() to possibly catch similar
issues earlier.

No functional change.

Closes #502
Closes #1074
master
1 parent e9f26cc commit 8b15961349e18a9ba113973c53f53913d0cd0fad
@joergoster joergoster committed