Rework the "unsupported" penalty into a "supported" bonus
Rework the "unsupported" penalty into a "supported" bonus
A pawn (according to all the searched positions of a bench run) is not
supported 85% of the time,
(in current master it is either isolated, backward or "unsupported").

So it made sense to try moving the S(17, 8) "unsupported" penalty value
into the base pawn value hoping for a more representative pawn value,
and accordingly
a) adjust backward and isolated so that they stay more or less the same
as master
b) increase the mg connected bonus in the supported case by S(17, 0) and
let the Connected formula find a suitable eg value according to rank.

Tested as a simplification SPRT(-3, 1)

Passed STC
http://tests.stockfishchess.org/tests/view/5970dbd30ebc5916ff649dd6
LLR: 2.95 (-2.94,2.94) [-3.00,1.00]
Total: 19613 W: 3663 L: 3540 D: 12410

Passed LTC
http://tests.stockfishchess.org/tests/view/597137780ebc5916ff649de3
LLR: 2.95 (-2.94,2.94) [-3.00,1.00]
Total: 24721 W: 3306 L: 3191 D: 18224

Bench: 5581946

Closes #1179
master (#445)
1 parent 722e1e0 commit b24bd762b2aff6cac30e1ae673775ea6f6aa8d11
@Rocky640 Rocky640 committed